article_system/pyproject.toml
[project]
name = "article-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.42",
]

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"
